requirements.txt
glfw==1.7.0
numpy==1.16.0
PyOpenGL==3.1.0
